Added setup.py
