+++ README.md
@@ -1,1 +1,9 @@
 # CTF Builder
+
+## Getting Started
+
+Install [Docker](https://docs.docker.com/get-docker/).
+
+```
+pip install .
+```
